fully functional solution

+++ 02_IncrTrainingLoad/definitions.hpp
@@ -1,4 +1,4 @@
 // short int
-typedef uint16_t sint;
+typedef uint32_t sint;
 // normal int
 typedef uint32_t nint;
